(#1146) WIP misc tidyups
