+++ src/hyperion/external_interaction/callbacks/ispyb_callback_base.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from abc import abstractmethod
 from typing import TYPE_CHECKING, Dict, Optional
 
 from hyperion.external_interaction.callbacks.plan_reactive_callback import (
@@ -110,8 +111,9 @@ class BaseISPyBCallback(PlanReactiveCallback):
             ISPYB_LOGGER.info(f"Recieved ISPYB IDs: {self.ispyb_ids}")
         return doc
 
+    @abstractmethod
     def update_deposition(self, params):
-        return self.ispyb.update_deposition(params)
+        pass
 
     def activity_gated_stop(self, doc: RunStop):
         """Subclasses must check that they are recieving a stop document for the correct
